Move the `Direction` enum to a separate file in helpers
