+++ src/config.rs
@@ -1,9 +1,6 @@
-use std::str::FromStr;
-use std::string::ParseError;
-
 use clap::Parser;
 
-use crate::helpers::colors;
+use crate::helpers::{colors, direction};
 use crate::symbols;
 
 //  =============
@@ -82,44 +79,5 @@ pub struct Config {
 
     /// The direction of motion for the stream particles
     #[clap(long, default_value = "down")]
-    pub direction: Direction,
-}
-
-#[derive(Default, Clone, Debug)]
-pub enum Direction {
-    #[default]
-    Down,
-    Up,
-    Left,
-    Right,
-}
-
-impl FromStr for Direction {
-    type Err = ParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "up" | "vertical-reverse" => Ok(Self::Up),
-            "down" | "vertical" => Ok(Self::Down),
-            "left" | "horizontal" => Ok(Self::Left),
-            "right" | "horizontal-reverse" => Ok(Self::Right),
-            _ => Ok(Self::default()),
-        }
-    }
-}
-
-impl Direction {
-    pub fn is_vertical(&self) -> bool {
-        match self {
-            Direction::Up | Direction::Down => true,
-            _ => false,
-        }
-    }
-
-    pub fn is_horizontal(&self) -> bool {
-        match self {
-            Direction::Left | Direction::Right => true,
-            _ => false,
-        }
-    }
+    pub direction: direction::Direction,
 }
