+++ src/helpers/mod.rs
@@ -1,2 +1,3 @@
 pub mod colors;
+pub mod direction;
 pub mod utils;
